cheat/sheet.py: Fix editing after copying a cheatsheet to DEFAULT_CHEAT_DIR
When editing (cheat -e) a unwritable cheatsheet, cheat offers to copy it
to DEFAULT_CHEAT_DIR and edit it there. However, due to the caching of
cheatsheets locations in sheets.cheat
(chrisallenlane/afcaaafbe59c0c24c41ec481a0b0b5cd913fa10f), sheet.edit()
uses the cached location of the sheet instead of the new location, hence
firing up the editor on the unwritable cheatsheet.

This is fixed by directly firing up the editor on the new location of
the cheatsheet.