Added winget installation instructions for Windows
Added installation instructions to include instructions to install cheat with winget. I submitted the winget manifest to be added as a winget package. As of writing this pull request, it needs to be verified by a moderator for it to be merged. If the manifest gets merged, then this pull request should be ready to merge.

Link to pull request: https://github.com/microsoft/winget-pkgs/pull/158530